Disable auto run when the query is going to alter the file system or the database
This important to avoid doing misstake on the file system.

It's rely on HTMX headers to inform the user and automaticaly disable
the auto run.

The database module is reorganize into multiple submodule for clarity.
A query object is introduce to later deal with stuff like syntax check.

The info/warning/error box are not dedicated component that can be
reuse.

The implementation for the detection is naive but it's the set the
foundation. It's never going to be perfect